move SkyDockerCommandRunner to skylet/providers
