+++ sky/skylet/providers/azure/node_provider.py
@@ -15,7 +15,7 @@ from sky.skylet.providers.azure.config import (
     bootstrap_azure,
     get_azure_sdk_function,
 )
-from sky.backends.docker_utils import SkyDockerCommandRunner
+from sky.skylet.providers.command_runner import SkyDockerCommandRunner
 from ray.autoscaler._private.command_runner import SSHCommandRunner
 from ray.autoscaler.node_provider import NodeProvider
 from ray.autoscaler.tags import (
